pyproject.toml
[tool.poetry]
name = "my-habits"
version = "0.1.0"
description = ""
authors = ["gelena <helenatcitok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
celery = "^5.4.0"
redis = "^5.0.4"
django-celery-beat = "^2.6.0"
requests = "^2.32.1"
coverage = "^7.5.1"
django-cors-headers = "^4.3.1"
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
